% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_cooperative.R
\name{opt_game}
\alias{opt_game}
\title{Cost Allocations based on the Optimistic Game for MCSTP}
\usage{
opt_game(C, sol = "shapley", draw = FALSE, which = c(1, 2), titles = TRUE)
}
\arguments{
\item{C}{a symmetric square matrix or a numeric vector representing the
lower triangle of costs (ordered by columns) among the nodes in \eqn{N_0}.
The first row and column are assumed to be the source (0). Supports \code{Inf}
for disconnected nodes.}

\item{sol}{character; the solution concept to apply to the game.
Currently, only \code{sol = "shapley"} (default) is supported.}

\item{draw}{logical; if \code{TRUE}, plots the network highlighting an optimal
tree in red (constructed using Prim's algorithm), indicating the cost allocated to
each agent in brackets below the node. For \eqn{n \le 3}, it also displays the
optimal network for each subcoalition \eqn{S \subset N}.}

\item{which}{numeric vector indicating which plots to display. If \code{1} (only
for \eqn{n \le 3}), displays the optimal network for each subcoalition
\eqn{S \subset N}; if \code{2}, the allocation for \eqn{N} is plotted.
Default is \code{c(1, 2)}.}

\item{titles}{logical; if \code{TRUE} (default), adds a main title
specifying the cooperative game and a subtitle with the solution concept used
and the total network cost.}
}
\value{
A list containing:
\itemize{
   \item \code{coalitions}: a character vector representing all possible coalitions \eqn{S \subseteq N} ordered by cardinality.
   \item \code{C_opt}: a list of the optimistic cost matrices \eqn{C^T} for each coalition \eqn{S \subseteq N}.
   \item \code{v_o}: a vector containing the characteristic function values \eqn{v^o(S)} for all coalitions.
   \item \code{contributions}: a data frame with the marginal contributions for all permutations \eqn{\pi \in \Pi_N}.
   \item \code{shapley}: the Shapley value allocation vector.
   \item \code{total}: the total cost of the MCST, \eqn{v^o(N)}.
   \item \code{nperms}: the number of permutations \eqn{n!} used to compute marginal contributions and the resulting Shapley value.
   \item \code{percentage}: the share of the total cost allocated to each agent.
   \item \code{ranking}: ranking of agents by cost (from highest to lowest; ties marked with *).
}
}
\description{
This function computes the cost allocation associated with the optimistic game
(Bergantiños and Vidal-Puga, 2007b) for a minimum cost spanning tree problem
\eqn{(N_0, C)}. The optimistic game, denoted as \eqn{(N, v_C^o)} \eqn{(}or simply
\eqn{(N, v^o)}\eqn{)}, evaluates the cost of each coalition \eqn{S \subseteq N}
assuming that agents in \eqn{N \setminus S} are already connected. Thus,
agents in \eqn{S} can connect to the source through agents in \eqn{N \setminus S}
for free. When the Shapley value is applied as the solution concept to this game,
the resulting allocation coincides with the Shapley value of the irreducible game \eqn{(N, v_C^i)}
(Bergantiños and Vidal-Puga, 2007b).
}
\details{
Let \eqn{T := N \setminus S}. The optimistic game \eqn{(N, v_C^o)} is a
cooperative game with transferable utility (TU game) defined for each
coalition \eqn{S \subseteq N} as:

\deqn{v_C^o(S) = m(S_0, C^T),}

where \eqn{m(S_0, C^T)} is the cost of the minimum spanning tree of the
subproblem induced by the agents in \eqn{S} and the source, using the
modified cost matrix \eqn{C^T}. In this matrix, the cost between agents
in \eqn{S} remains the same, i.e., \eqn{c_{ij}^T = c_{ij}} for all \eqn{i,j \in S},
but the cost to connect any agent \eqn{i \in S} to the source is updated as:

\deqn{c_{0i}^T = \min_{j \in T \cup \{0\}} c_{ji}.}

Applying the Shapley value to this TU game assigns to each agent \eqn{i \in N}
the average of their marginal contributions over all possible permutations
\eqn{\pi \in \Pi_N}:

\deqn{Sh_i(N, v^o) = \displaystyle{\frac{1}{n!} \sum_{\pi \in \Pi_N} [v^o(Pre(i, \pi) \cup \{i\}) - v^o(Pre(i, \pi))]},}

where \eqn{Pre(i, \pi)} is the set of players that precede agent \eqn{i}
in the ordering \eqn{\pi}.
}
\note{
The optimistic game associated with any MCST problem coincides with the optimistic
game associated with its irreducible form, i.e.,

\deqn{v_C^o = v_{\bar{C}}^o.}

If \eqn{(N_0, \bar{C})} is irreducible, the private and optimistic games, \eqn{v^p}
and \eqn{v^o}, are dual: \eqn{v^p(S) + v^o(N \setminus S) = m(N_0, C)} for all \eqn{S \subset N}.

Furthermore, as \eqn{Sh(N, v_C^o) = Sh(N, v_C^i)}, applying the
Shapley value to \eqn{v^o} is simply another way of obtaining the
\code{\link{folk_rule}} (Bergantiños and Vidal-Puga, 2007b).
}
\examples{
# Simple vector input
opt_game(c(12, 15, 20, 4, 6, 8), sol = "shapley", draw = TRUE)

# Input with infinite costs (disconnected nodes)
C_inf <- c(5, 9, Inf, 15, 6, Inf, 7, Inf, Inf, Inf,
           Inf, 8, 7, Inf, Inf, 5, Inf, Inf, 8, 9, 11)
opt_game(C_inf, sol = "shapley")

# Matrix input
C_mat <- matrix(c(0, 12, 15, 12,
                 12,  0,  4,  6,
                 15,  4,  0,  8,
                 12,  6,  8,  0), byrow = TRUE, ncol = 4)
opt_game(C_mat, draw = TRUE, which = 2, titles = FALSE)

}
\references{
Bergantiños G, Vidal-Puga J (2007b) The optimistic TU game in minimum cost
spanning tree problems. Int J Game Theory 36(2):223–239

Bergantiños G, Vidal-Puga J (2021) A review of cooperative rules
and their associated algorithms for minimum-cost spanning tree problems.
SERIEs, 12:73-100
}
\seealso{
\code{\link{private_game}} for a game based on a pessimistic approach.

\code{\link{irred_game}} for an equivalent game under the Shapley value.

\code{\link{alloc_rules}} for an overview of the available rules in the package.
}
\concept{Cooperative Games}
\concept{MCSTP}
