% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcstprules-package.R
\docType{package}
\name{mcstprules-package}
\alias{mcstprules}
\alias{mcstprules-package}
\title{mcstprules: Cost Allocation Rules in Minimum Cost Spanning Tree Problems}
\description{
A comprehensive set of tools to compute, analyze, and visualize cost allocation rules
for Minimum Cost Spanning Tree Problems (MCSTP).
}
\details{
This package implements two major approaches for cost allocation found in the literature:
algorithmic rules and rules defined through cooperative games. Additionally, it provides
a robust suite of analytical tools to evaluate core stability, geometric properties,
and cost sensitivity.

For a detailed overview, classification, and references of all
implemented rules and analysis tools, please see the central documentation page:
\code{\link{mcstRules}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luciasouto13/mcstprules}
  \item \url{https://luciasouto13.github.io/mcstprules/}
  \item Report bugs at \url{https://github.com/luciasouto13/mcstprules/issues}
}

}
\author{
\strong{Maintainer}: Lucía Souto Vázquez \email{luciasouvaz@gmail.com}

}
\keyword{internal}
